moved ui code to it's own module
- This commit does not fix any coding erros, just moved the code around to
make reviewing easier.
- The code was not deduplicated.
- The code used for the webui was moved from the api module to the new ui
module, and renamed to static.
- Fixed the paths for the static content
- Removed the patches module.
